Update the goo.im JSON endpoint
Recently, goo.im pushed an update for their JSON API. The location of
the endpoint used by the application was moved around (making SSL a
requirement) and the output returned by the call also got changed.

It should be noted that this commit does not make downloads of GApps
packages possible as the CDN requires a token which changes all the
time and is generated using a private key (held by goo.im) instead
aiming to get rid of the pesky error toast notifying about the update
checker failure.