Skip downloads directory initialization
It is safe to do this clean-up since the method being called will do
all the work automatically anyway if it is in a bad state. A bad state
here is the directory not being created. The app will encounter a
larger issue if the downloads directory cannot be created in which
case a runtime exception will bubble up but the initial call won't
help out with that anyway.