Rename the dictionary variable and close IOUtils
This is a simple change that won't really affect anything. This also
kind of marks the point at which IOUtils is being closed, at least
for now. Some cleanup is still needed here - on the primary/secondary
external storage code and mountpoint checking, but that can be left
to be for now.